requirements.txt
asgiref==3.11.1
cffi==2.0.0
crispy-bootstrap5==2025.6
cryptography==46.0.5
Django==6.0.2
django-appconf==1.2.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-cryptography==1.1
django-secured-fields==0.4.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pycparser==3.0
PyJWT==2.11.0
sqlparse==0.5.5
